+++ router-tests/testenv/testenv.go
@@ -3,11 +3,14 @@ package testenv
 import (
 	"bytes"
 	"context"
+	"crypto/tls"
+	"crypto/x509"
 	_ "embed"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
+	"log"
 	"math/rand"
 	"net/http"
 	"net/http/httptest"
@@ -18,6 +21,8 @@ import (
 	"sync"
 	"testing"
 	"time"
+
+	"github.com/hashicorp/go-cleanhttp"
 
 	"github.com/golang-jwt/jwt/v5"
 	"github.com/gorilla/websocket"
@@ -74,6 +79,7 @@ type Config struct {
 	ModifyEngineExecutionConfiguration func(engineExecutionConfiguration *config.EngineExecutionConfiguration)
 	ModifyCDNConfig                    func(cdnConfig *config.CDNConfiguration)
 	DisableWebSockets                  bool
+	TLSConfig                          *core.TlsConfig
 }
 
 type SubgraphsConfig struct {
@@ -266,10 +272,11 @@ func createTestEnv(t testing.TB, cfg *Config) (*Environment, error) {
 	}
 
 	listenerAddr := fmt.Sprintf("localhost:%d", routerPort)
-	routerURL := fmt.Sprintf("http://%s", listenerAddr)
 
 	client := retryablehttp.NewClient()
 	client.Logger = nil
+	client.RetryMax = 10
+	client.RetryWaitMin = 100 * time.Millisecond
 
 	rr, err := configureRouter(listenerAddr, cfg, &routerConfig, cdn, ns)
 	if err != nil {
@@ -279,9 +286,40 @@ func createTestEnv(t testing.TB, cfg *Config) (*Environment, error) {
 	svr, err := rr.NewServer(ctx)
 	require.NoError(t, err)
 
+	if cfg.TLSConfig != nil && cfg.TLSConfig.Enabled {
+
+		cert, err := tls.LoadX509KeyPair(cfg.TLSConfig.CertFile, cfg.TLSConfig.KeyFile)
+		require.NoError(t, err)
+
+		caCert, err := os.ReadFile(cfg.TLSConfig.CertFile)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		caCertPool := x509.NewCertPool()
+		if ok := caCertPool.AppendCertsFromPEM(caCert); !ok {
+			t.Fatalf("could not append ca cert to pool")
+		}
+
+		// Retain the default transport settings
+		httpClient := cleanhttp.DefaultPooledClient()
+		httpClient.Transport.(*http.Transport).TLSClientConfig = &tls.Config{
+			RootCAs:      caCertPool,
+			Certificates: []tls.Certificate{cert},
+		}
+
+		client.HTTPClient = httpClient
+	}
+
 	go func() {
-		if err := svr.HttpServer().ListenAndServe(); err != nil && !errors.Is(err, http.ErrServerClosed) {
-			t.Errorf("could not start router: %s", err)
+		if cfg.TLSConfig != nil && cfg.TLSConfig.Enabled {
+			if err := svr.HttpServer().ListenAndServeTLS("", ""); err != nil && !errors.Is(err, http.ErrServerClosed) {
+				t.Errorf("could not start tls router: %s", err)
+			}
+		} else {
+			if err := svr.HttpServer().ListenAndServe(); err != nil && !errors.Is(err, http.ErrServerClosed) {
+				t.Errorf("could not start router: %s", err)
+			}
 		}
 	}()
 
@@ -327,7 +365,7 @@ func createTestEnv(t testing.TB, cfg *Config) (*Environment, error) {
 		Context:              ctx,
 		cancel:               cancel,
 		Router:               rr,
-		RouterURL:            routerURL,
+		RouterURL:            svr.BaseURL(),
 		RouterClient:         client.StandardClient(),
 		CDN:                  cdn,
 		Nats:                 ns,
@@ -344,7 +382,9 @@ func createTestEnv(t testing.TB, cfg *Config) (*Environment, error) {
 			countriesServer,
 		},
 	}
+
 	e.waitForRouterConnection(ctx)
+
 	return e, nil
 }
 
@@ -408,6 +448,7 @@ func configureRouter(listenerAddr string, testConfig *Config, routerConfig *node
 		core.WithEngineExecutionConfig(engineExecutionConfig),
 		core.WithCDN(cfg.CDN),
 		core.WithListenerAddr(listenerAddr),
+		core.WithTLSConfig(testConfig.TLSConfig),
 		core.WithEvents(config.EventsConfiguration{
 			Sources: []config.EventSource{
 				{
@@ -544,20 +585,16 @@ type GraphQLRequest struct {
 type TestResponse struct {
 	Body     string
 	Response *http.Response
+	Proto    string
 }
 
 func (e *Environment) waitForRouterConnection(ctx context.Context) {
-	// Wait for the router to be ready
-	client := &http.Client{
-		Transport: http.DefaultTransport,
-		Timeout:   time.Millisecond * 100,
-	}
 	for {
 		select {
 		case <-ctx.Done():
 			e.t.Fatalf("timed out waiting for router to be ready")
 		default:
-			resp, err := client.Get(e.RouterURL)
+			resp, err := e.RouterClient.Get(e.RouterURL)
 			if err == nil {
 				_ = resp.Body.Close()
 				return
@@ -584,11 +621,7 @@ func (e *Environment) MakeGraphQLRequest(request GraphQLRequest) (*TestResponse,
 	if request.Header != nil {
 		req.Header = request.Header
 	}
-	client := &http.Client{
-		Transport: http.DefaultTransport,
-		Timeout:   time.Second * 5,
-	}
-	resp, err := client.Do(req)
+	resp, err := e.RouterClient.Do(req)
 	if err != nil {
 		return nil, err
 	}
@@ -603,6 +636,7 @@ func (e *Environment) MakeGraphQLRequest(request GraphQLRequest) (*TestResponse,
 	return &TestResponse{
 		Body:     strings.TrimSpace(body),
 		Response: resp,
+		Proto:    resp.Proto,
 	}, nil
 }
 
@@ -611,17 +645,14 @@ func (e *Environment) MakeRequest(method, path string, header http.Header, body 
 	if err != nil {
 		return nil, err
 	}
+
 	req, err := http.NewRequestWithContext(e.Context, method, requestURL, body)
 	if err != nil {
 		return nil, err
 	}
 	req.Header = header
-	client := &http.Client{
-		Transport: http.DefaultTransport,
-		Timeout:   time.Second * 5,
-	}
-	return client.Do(req)
 
+	return e.RouterClient.Do(req)
 }
 
 func (e *Environment) GraphQLRequestURL() string {
@@ -893,6 +924,37 @@ func (e *Environment) WaitForMessagesSent(desiredCount uint64, timeout time.Dura
 			}
 		}
 	}
+}
+
+func (e *Environment) WaitForTriggerCount(desiredCount uint64, timeout time.Duration) {
+	e.t.Helper()
+
+	report := e.Router.WebsocketStats.GetReport()
+	if report.Triggers == desiredCount {
+		return
+	}
+
+	ctx, cancel := context.WithTimeout(e.Context, timeout)
+	defer cancel()
+
+	sub := e.Router.WebsocketStats.Subscribe(ctx)
+
+	for {
+		select {
+		case <-ctx.Done():
+			e.t.Fatalf("timed out waiting for trigger count, got %d, want %d", report.Triggers, desiredCount)
+			return
+		case report, ok := <-sub:
+			if !ok {
+				e.t.Fatalf("timed out waiting for trigger count, got %d, want %d", report.Triggers, desiredCount)
+				return
+			}
+			if report.Triggers == desiredCount {
+				return
+			}
+		}
+	}
+
 }
 
 func subgraphOptions(t testing.TB, ns *natsserver.Server) *subgraphs.SubgraphOptions {